r2223: The chdesc history command "status <chdesc>" did not work for chdescs with an address using the high bit of 32bits because Java's Integer.parseInt() does will not parse such large numbers. Updated SystemState.unhex() to instead use Long.parseLong() and cast the result down to an int. Code that goes in the other direction (int to String) is fine.

+++ util/debug/SystemState.java
@@ -198,7 +198,7 @@ public class SystemState
 	public static int unhex(String hex)
 	{
 		if(!hex.substring(0, 2).equals("0x"))
-			return Integer.parseInt(hex);
-		return Integer.parseInt(hex.substring(2), 16);
+			return (int) Long.parseLong(hex);
+		return (int) Long.parseLong(hex.substring(2), 16);
 	}
 }
